refactor: extract benchmark catalog types
