+++ devtools/quality_registry.py
@@ -4,18 +4,18 @@ from __future__ import annotations
 
 from dataclasses import dataclass
 
-from devtools.benchmark_campaign import CAMPAIGNS as BENCHMARK_CAMPAIGNS
-from devtools.benchmark_campaigns import SYNTHETIC_BENCHMARK_SCENARIOS
+from devtools.benchmark_catalog import (
+    BenchmarkCampaignEntry,
+    build_benchmark_entries,
+    build_synthetic_benchmark_entries,
+)
 from devtools.mutmut_campaign import CAMPAIGNS as MUTATION_CAMPAIGNS
 from devtools.scenario_projection_catalog import build_scenario_projection_entries
 from devtools.validation_lane_base import LaneConfig
 from devtools.validation_lane_catalog_composites import COMPOSITE_LANES
 from devtools.validation_lane_catalog_contracts import CONTRACT_LANES
 from devtools.validation_lane_catalog_live import LIVE_LANES
-from polylogue.scenarios import (
-    ScenarioMetadata,
-    ScenarioProjectionEntry,
-)
+from polylogue.scenarios import ScenarioProjectionEntry
 
 
 @dataclass(frozen=True)
@@ -34,16 +34,6 @@ class MutationCampaignEntry:
     paths_to_mutate: tuple[str, ...]
     tests: tuple[str, ...]
     notes: tuple[str, ...] = ()
-
-
-@dataclass(frozen=True)
-class BenchmarkCampaignEntry(ScenarioMetadata):
-    name: str
-    description: str
-    tests: tuple[str, ...]
-    notes: tuple[str, ...] = ()
-    warn_pct: float = 0.0
-    fail_pct: float = 0.0
 
 
 @dataclass(frozen=True)
@@ -85,58 +75,19 @@ def _mutation_entries() -> tuple[MutationCampaignEntry, ...]:
     return tuple(sorted(entries, key=lambda item: item.name))
 
 
-def _benchmark_entries() -> tuple[BenchmarkCampaignEntry, ...]:
-    entries = [
-        BenchmarkCampaignEntry(
-            name=campaign.name,
-            description=campaign.description,
-            tests=tuple(campaign.tests),
-            notes=tuple(campaign.notes),
-            warn_pct=campaign.warn_pct,
-            fail_pct=campaign.fail_pct,
-            origin=campaign.origin,
-            path_targets=tuple(campaign.path_targets),
-            artifact_targets=tuple(campaign.artifact_targets),
-            operation_targets=tuple(campaign.operation_targets),
-            tags=tuple(campaign.tags),
-        )
-        for campaign in BENCHMARK_CAMPAIGNS.values()
-    ]
-    return tuple(sorted(entries, key=lambda item: item.name))
-
-
-def _synthetic_benchmark_entries() -> tuple[BenchmarkCampaignEntry, ...]:
-    entries = [
-        BenchmarkCampaignEntry(
-            name=scenario.scenario_id,
-            description=scenario.description,
-            tests=(),
-            notes=(),
-            origin=scenario.origin,
-            path_targets=scenario.path_targets,
-            artifact_targets=scenario.artifact_targets,
-            operation_targets=scenario.operation_targets,
-            tags=scenario.tags,
-        )
-        for scenario in SYNTHETIC_BENCHMARK_SCENARIOS
-    ]
-    return tuple(sorted(entries, key=lambda item: item.name))
-
-
 def build_quality_registry() -> QualityRegistry:
     return QualityRegistry(
         contract_lanes=_lane_entries("contract", CONTRACT_LANES),
         live_lanes=_lane_entries("live", LIVE_LANES),
         composite_lanes=_lane_entries("composite", COMPOSITE_LANES),
         mutation_campaigns=_mutation_entries(),
-        benchmark_campaigns=_benchmark_entries(),
-        synthetic_benchmark_campaigns=_synthetic_benchmark_entries(),
+        benchmark_campaigns=build_benchmark_entries(),
+        synthetic_benchmark_campaigns=build_synthetic_benchmark_entries(),
         scenario_projections=build_scenario_projection_entries(),
     )
 
 
 __all__ = [
-    "BenchmarkCampaignEntry",
     "MutationCampaignEntry",
     "QualityRegistry",
     "ValidationLaneEntry",
